requirements.txt
python-telegram-bot==21.6
PyMySQL==1.1.1
openpyxl==3.1.5